Seen episodes remembered.
Now the user can mark episodes as seen and the program doesn't forget.
The Series class has a SeenEpisodes list which holds the seen episodes.
Then the function MarkSeenEpisodes() does the job of updating the actual
Episodes list IsSeen boolean by comparing with SeasonNumber and
EpisodeNumber of each SeenEpisode in the list.
The MarkSeenEpisodes() is called in multiple spots to ensure that it'll
work, I'm not very familiar with the code yet so maybe you can have a
look and remove it from the unnecessary places.

Bugs to note:
1) Sometimes when I don't setup the connectivity of my emulator and it
doesn't have internet, instead of handling it nicely it crashes. Would
this happen to the app in case connectivity is lost for whatever reason?
2) When a series is selected in the 'Followed' list, the name text
disappears (until deselected).